# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: FoodFlow API
  description: |
    FoodFlow -- Food Delivery Platform REST API.

    **Base path:** `/api`. The production origin is the verified Vercel API alias.

    **Conventions:**
    - All monetary values are in **VND**, represented as integers.
    - All datetime values use ISO 8601 `date-time` with timezone.
    - Successful web/admin/restaurant endpoints return `{ success: true, data, meta? }`.
    - List endpoints put the collection in `data` and pagination in `meta`.
    - Errors use the shared `ProblemDetail` schema (RFC 7807) and are not success-wrapped.
    - Idempotency is supported via `Idempotency-Key` header (UUID v4).
    - Inbound HMAC: SePay webhooks are verified with `x-sepay-signature`.
    - Outbound HMAC: generic outbound webhooks use `X-Signature-SHA256` when configured.

    **Versioning:** Batch 4 currently uses the `/api` base without a `/v1` prefix. A future version prefix requires an updated server contract and migration policy.
  version: "1.0.0"
  contact:
    name: FoodFlow Team
    email: support@foodflow.vn
servers:
  - url: https://replace-with-foodflow-api.vercel.app/api
    description: Production placeholder; replace with the verified Vercel API alias before publishing external client configuration
  - url: http://localhost:3001/api
    description: Local development
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer access token. Obtain via `POST /auth/login`; the signing algorithm follows the active JWT cutover phase.
    cronBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Exact server-side CRON_SECRET value. Never expose this credential to browser or mobile clients.
...
  schemas:
...
    ProblemDetail:
      type: object
      properties:
        type:    { type: string, description: "URI to problem type documentation" }
        title:   { type: string, description: "Short human-readable summary" }
        status:  { type: integer, description: "HTTP status code" }
        detail:  { type: string, description: "Human-readable explanation" }
        instance: { type: string, description: "URI to the specific occurrence" }
        code:    { type: string, description: "Application-specific error code (e.g. AUTH_INVALID_CREDENTIALS)" }
      required: [type, title, status, detail]
      example:
        type: "about:blank"
        title: "Validation Error"
        status: 422
        detail: "Email is not valid"
        code: "VALIDATION_INVALID_EMAIL"
...
    OrderTrackingDriverLocation:
      type: object
      description: Last real GPS position recorded for the driver currently assigned to the order.
      properties:
        lat: { type: number, minimum: -90, maximum: 90 }
        lng: { type: number, minimum: -180, maximum: 180 }
        lastUpdated: { type: string, format: date-time }
      required: [lat, lng, lastUpdated]
    OrderTrackingResponse:
      type: object
      description: Order-participant-scoped tracking snapshot. Nullable fields remain null until real telemetry exists.
      properties:
        orderId: { type: string, format: uuid }
        status: { type: string, enum: [created, pending_payment, paid, restaurant_pending, restaurant_accepted, preparing, ready_for_pickup, driver_assigned, driver_arriving_restaurant, picked_up, delivering, delivered, completed, cancelled, refunded] }
        driverLocation:
          oneOf:
            - $ref: "#/components/schemas/OrderTrackingDriverLocation"
            - type: "null"
        etaMinutes: { type: [integer, "null"], minimum: 1 }
        routePolyline: { type: [string, "null"] }
        routePhase: { type: string, enum: [pickup, dropoff] }
      required: [orderId, status, driverLocation, etaMinutes, routePolyline, routePhase]
...
    GoOnlineRequest:
      type: object
      properties:
        lat: { type: number, minimum: -90, maximum: 90 }
        lng: { type: number, minimum: -180, maximum: 180 }
      required: [lat, lng]
      example:
        lat: 10.7769
        lng: 106.7009
    DriverLocationUpdateRequest:
      type: object
      additionalProperties: false
      description: A fresh GPS sample captured by the authenticated driver's device. Speed is expressed in kilometres per hour.
      properties:
        lat: { type: number, minimum: -90, maximum: 90 }
        lng: { type: number, minimum: -180, maximum: 180 }
        bearing: { type: number, minimum: 0, exclusiveMaximum: 360 }
        speed: { type: number, minimum: 0, maximum: 150 }
        accuracy: { type: number, minimum: 0, maximum: 10000 }
        timestamp: { type: string, format: date-time }
      required: [lat, lng, timestamp]
    DriverLocationUpdateResponse:
      type: object
      properties:
        accepted: { type: boolean, const: true }
        orderId: { type: [string, "null"], format: uuid }
        timestamp: { type: string, format: date-time }
      required: [accepted, orderId, timestamp]
...
    DriverEarningEntry:
      type: object
      description: One real payout ledger row joined to its completed order.
      properties:
        orderId: { type: string, format: uuid }
        orderCode: { type: string }
        restaurantName: { type: string }
        amount: { type: integer, description: "Net driver payout ledger amount in VND" }
        completedAt: { type: string, format: date-time }
      required: [orderId, orderCode, restaurantName, amount, completedAt]
    DriverEarnings:
      type: object
      description: Driver payout ledger totals for the requested `today`, `week`, or `month` period.
      properties:
        totalEarnings: { type: integer, description: "Net driver payout ledger amount in VND" }
        totalOrders: { type: integer, minimum: 0, description: "Distinct orders represented by the payout entries." }
        averagePerOrder: { type: integer, description: "Rounded VND average per distinct order." }
        entries:
          type: array
          items: { $ref: "#/components/schemas/DriverEarningEntry" }
      required: [totalEarnings, totalOrders, averagePerOrder, entries]
      example:
        totalEarnings: 350000
        totalOrders: 7
        averagePerOrder: 50000
        entries:
          - orderId: "66666666-6666-4666-8666-666666666666"
            orderCode: "FF-20260703-0001"
            restaurantName: "Pho 24"
            amount: 50000
            completedAt: "2026-07-03T08:30:00Z"
    DriverOrderItem:
      type: object
      properties:
        menuItemId: { type: string, format: uuid }
        name: { type: string }
        quantity: { type: integer, minimum: 1 }
        unitPrice: { type: integer, description: "VND" }
        totalPrice: { type: integer, description: "VND" }
        selectedOptions:
          description: Original selected option JSON persisted on the order item.
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items: {}
            - type: string
            - type: number
            - type: boolean
            - type: "null"
      required: [menuItemId, name, quantity, unitPrice, totalPrice, selectedOptions]
    DriverOrderAddress:
      type: object
      properties:
        label: { type: string }
        address: { type: string }
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
      required: [label, address, latitude, longitude]
    DriverOrderStatusHistory:
      type: object
      properties:
        status: { type: string }
        timestamp: { type: string, format: date-time }
        note: { type: [string, "null"] }
      required: [status, timestamp, note]
    DriverOrder:
      type: object
      description: Order snapshot scoped to the authenticated driver. Coordinates are read from PostGIS; missing location rows return `404 ORDER_LOCATION_NOT_FOUND` instead of synthetic coordinates.
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        restaurantId: { type: string, format: uuid }
        restaurantName: { type: string }
        restaurantLogoUrl: { type: [string, "null"] }
        restaurantPhone: { type: [string, "null"] }
        customerName: { type: string }
        customerPhone: { type: [string, "null"] }
        items:
          type: array
          items: { $ref: "#/components/schemas/DriverOrderItem" }
        subtotal: { type: integer, description: "VND" }
        deliveryFee: { type: integer, description: "VND" }
        discount: { type: integer, description: "VND" }
        total: { type: integer, description: "VND" }
        status: { type: string, enum: [driver_assigned, driver_arriving_restaurant, picked_up, delivering, delivered, completed, cancelled, refunded] }
        driverId: { type: [string, "null"], format: uuid }
        deliveryAddress: { $ref: "#/components/schemas/DriverOrderAddress" }
        paymentMethod: { type: string }
        note: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        statusHistory:
          type: array
          items: { $ref: "#/components/schemas/DriverOrderStatusHistory" }
        restaurantLatitude: { type: number, minimum: -90, maximum: 90 }
        restaurantLongitude: { type: number, minimum: -180, maximum: 180 }
        estimatedDeliveryTimeMinutes: { type: [integer, "null"], minimum: 1 }
        routePolyline: { type: [string, "null"] }
      required:
        - id
        - userId
        - restaurantId
        - restaurantName
        - restaurantLogoUrl
        - restaurantPhone
        - customerName
        - customerPhone
        - items
        - subtotal
        - deliveryFee
        - discount
        - total
        - status
        - driverId
        - deliveryAddress
        - paymentMethod
        - note
        - createdAt
        - updatedAt
        - statusHistory
        - restaurantLatitude
        - restaurantLongitude
        - estimatedDeliveryTimeMinutes
        - routePolyline
    DriverDailyEarning:
      type: object
      properties:
        date: { type: string, format: date }
        amount: { type: integer, description: "Net driver payout ledger amount in VND" }
        tripCount: { type: integer, minimum: 0 }
      required: [date, amount, tripCount]
    DriverEarningsSummary:
      type: object
      properties:
        period: { type: string, enum: [7d, 30d, 90d] }
        totalVnd: { type: integer, description: "Net driver payout ledger amount in VND" }
        tripCount: { type: integer, minimum: 0 }
        avgPerTrip: { type: integer, description: "VND" }
        byDay:
          type: array
          items: { $ref: "#/components/schemas/DriverDailyEarning" }
      required: [period, totalVnd, tripCount, avgPerTrip, byDay]
    DriverEarningsSummaryResponse:
      type: object
      properties:
        success: { type: boolean, const: true }
        data: { $ref: "#/components/schemas/DriverEarningsSummary" }
      required: [success, data]
    DriverRatingReview:
      type: object
      properties:
        id: { type: string, format: uuid }
        customerName: { type: string }
        customerAvatarUrl: { type: string, nullable: true }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string, nullable: true }
        date: { type: string, format: date-time }
        orderId: { type: string, format: uuid }
        orderCode: { type: string }
      required: [id, customerName, customerAvatarUrl, rating, comment, date, orderId, orderCode]
    DriverRatingStats:
      type: object
      properties:
        average: { type: number, format: double, minimum: 0, maximum: 5 }
        totalReviews: { type: integer, minimum: 0 }
        distribution:
          type: object
          additionalProperties:
            type: integer
            minimum: 0
          example: { "1": 0, "2": 0, "3": 0, "4": 1, "5": 2 }
      required: [average, totalReviews, distribution]
    DriverRatings:
      type: object
      properties:
        reviews:
          type: array
          items: { $ref: "#/components/schemas/DriverRatingReview" }
        stats: { $ref: "#/components/schemas/DriverRatingStats" }
        hasMore: { type: boolean }
      required: [reviews, stats, hasMore]
    DriverRatingsResponse:
      type: object
      properties:
        success: { type: boolean, const: true }
        data: { $ref: "#/components/schemas/DriverRatings" }
      required: [success, data]
    DriverTripRoutePoint:
      type: object
      properties:
        lat: { type: number, format: double }
        lng: { type: number, format: double }
        timestamp: { type: string, format: date-time }
      required: [lat, lng, timestamp]
    DriverTripRouteSegment:
      type: object
      properties:
        distanceKm: { type: number, format: double, minimum: 0 }
        durationSeconds: { type: integer, minimum: 0 }
        instruction: { type: string }
        startIndex: { type: integer, minimum: 0 }
        endIndex: { type: integer, minimum: 0 }
      required: [distanceKm, durationSeconds, instruction, startIndex, endIndex]
    DriverTripRoute:
      type: object
      properties:
        tripId: { type: string, format: uuid }
        points:
          type: array
          items: { $ref: "#/components/schemas/DriverTripRoutePoint" }
        segments:
          type: array
          items: { $ref: "#/components/schemas/DriverTripRouteSegment" }
        totalDistanceKm: { type: number, format: double, minimum: 0 }
        totalDurationSeconds: { type: integer, minimum: 0 }
        avgSpeedKmh: { type: number, format: double, minimum: 0 }
        payout: { type: number, minimum: 0 }
      required: [tripId, points, segments, totalDistanceKm, totalDurationSeconds, avgSpeedKmh, payout]
    DriverTripRouteResponse:
      type: object
      properties:
        success: { type: boolean, const: true }
        data: { $ref: "#/components/schemas/DriverTripRoute" }
      required: [success, data]
    DriverHeatmapPoint:
      type: object
      properties:
        lat: { type: number, format: double }
        lng: { type: number, format: double }
        demandLevel: { type: integer, minimum: 0, maximum: 2, description: "0 low, 1 medium, 2 high" }
        orderCount: { type: integer, minimum: 0 }
        avgPayout: { type: integer, description: "Average delivery fee in VND" }
      required: [lat, lng, demandLevel, orderCount, avgPayout]
      example:
        lat: 10.7769
        lng: 106.7009
        demandLevel: 2
        orderCount: 12
        avgPayout: 32000
    DriverHeatmapResponse:
      type: object
      properties:
        success: { type: boolean, const: true }
        data:
          type: array
          items: { $ref: "#/components/schemas/DriverHeatmapPoint" }
      required: [success, data]
...
paths:
...
  /orders/{orderId}/tracking:
    get:
      tags: [orders]
      summary: Get the current order tracking snapshot
      description: Returns an accessible order's status plus real assigned-driver GPS, cached or persisted ETA, and route data. Customers, assigned drivers, active restaurant staff for the order tenant, and admins may read the snapshot. Telemetry fields remain null until recorded and are never replaced with synthetic coordinates.
      operationId: orders_getTracking
      x-foo-actor: user
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: "Current tracking snapshot", content: { application/json: { schema: { $ref: "#/components/schemas/OrderTrackingResponse" } } } }
        "401": { description: "Authentication required", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
        "403": { description: "Order participant role required", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
        "404": { description: "Order not found for this actor", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
...
  /driver/online:
    post:
      tags: [drivers]
      summary: Go online (start accepting orders)
      description: Marks the authenticated driver online and records their latest service location.
      operationId: drivers_goOnline
      x-foo-actor: driver
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GoOnlineRequest" }
      responses:
        "200": { description: "Now online", content: { application/json: { schema: { type: object, properties: { isOnline: { type: boolean }, lat: { type: number }, lng: { type: number } } } } } }
...
  /driver/location:
    post:
      tags: [drivers]
      summary: Submit a live driver GPS sample
      description: Reuses the authenticated tracking pipeline and rejects stale, future, out-of-bounds, implausibly fast, or teleporting samples before updating Redis/PostGIS and publishing authorized realtime order/admin events.
      operationId: tracking_updateDriverLocation
      x-foo-actor: driver
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DriverLocationUpdateRequest" }
      responses:
        "201":
          description: Location sample accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DriverLocationUpdateResponse" }
        "422":
          description: Location rejected as stale, future, invalid, outside the service area, or physically implausible
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProblemDetail" }
...
  /driver/orders/active:
    get:
      tags: [drivers]
      summary: Get authenticated driver's active order
      description: Returns the newest active order assigned to the authenticated driver, or `null` when no active assignment exists. Order and address coordinates come from PostGIS; missing location rows return `404 ORDER_LOCATION_NOT_FOUND`.
      operationId: drivers_getActiveOrder
      x-foo-actor: driver
      responses:
        "200":
          description: Active order or null
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/DriverOrder"
                  - type: "null"
        "404": { description: "Order location not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
  /driver/orders/history:
    get:
      tags: [drivers]
      summary: List authenticated driver's completed order history
      description: Returns terminal orders assigned to the authenticated driver. Empty arrays are valid; API failures must not be treated as empty history by clients.
      operationId: drivers_getOrderHistory
      x-foo-actor: driver
      parameters:
        - name: fromDate
          in: query
          schema: { type: string, format: date-time }
        - name: toDate
          in: query
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        "200":
          description: Driver order history
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DriverOrder" }
        "404": { description: "Order location not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
...
  /driver/earnings:
    get:
      tags: [drivers]
      summary: Get driver payout ledger earnings
      description: Returns payout ledger totals and entries for the authenticated driver over the requested `today`, `week`, or `month` period. Empty `entries` means no settled driver payout rows exist for the period.
      operationId: drivers_getEarnings
      x-foo-actor: driver
      parameters:
        - name: period
          in: query
          schema: { type: string, enum: [today, week, month], default: today }
      responses:
        "200": { description: "Earnings", content: { application/json: { schema: { $ref: "#/components/schemas/DriverEarnings" } } } }
        "400": { description: "Invalid earnings period", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
  /driver/earnings/summary:
    get:
      tags: [drivers]
      summary: Get chart-ready driver earnings summary
      description: Returns continuous day buckets from driver payout ledger data. Days without trips are returned with zero amounts.
      operationId: drivers_getEarningsSummary
      x-foo-actor: driver
      parameters:
        - name: period
          in: query
          schema: { type: string, enum: [7d, 30d, 90d], default: 7d }
      responses:
        "200": { description: "Earnings summary", content: { application/json: { schema: { $ref: "#/components/schemas/DriverEarningsSummaryResponse" } } } }
  /driver/ratings:
    get:
      tags: [drivers]
      summary: Get driver rating history
      description: Returns visible delivery reviews for the authenticated driver and derives rating stats from Review rows. Empty arrays are valid; no runtime sample reviews are generated.
      operationId: drivers_getRatings
      x-foo-actor: driver
      parameters:
        - name: star
          in: query
          schema: { type: integer, minimum: 1, maximum: 5 }
      responses:
        "200": { description: "Rating history", content: { application/json: { schema: { $ref: "#/components/schemas/DriverRatingsResponse" } } } }
  /driver/trips/{tripId}/route:
    get:
      tags: [drivers]
      summary: Get driver trip route
      description: Returns real telemetry or persisted route geometry for a trip owned by the authenticated driver. Empty arrays are valid when no route data has been recorded; no runtime sample route is generated.
      operationId: drivers_getTripRoute
      x-foo-actor: driver
      parameters:
        - name: tripId
          in: path
          required: true
          schema: { type: string }
          description: Order UUID or order code.
      responses:
        "200": { description: "Trip route", content: { application/json: { schema: { $ref: "#/components/schemas/DriverTripRouteResponse" } } } }
        "404": { description: "Trip not found for driver", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
  /driver/heatmap:
    get:
      tags: [drivers]
      summary: Get nearby driver demand heatmap
      description: Returns real order-density points near the authenticated driver's current map center. Empty arrays are valid; no runtime sample data is generated.
      operationId: drivers_getHeatmap
      x-foo-actor: driver
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number, minimum: -90, maximum: 90 }
        - name: lng
          in: query
          required: true
          schema: { type: number, minimum: -180, maximum: 180 }
        - name: radius
          in: query
          schema: { type: number, minimum: 0.1, maximum: 25, default: 5 }
        - name: window
          in: query
          schema: { type: string, enum: [now, 1h, 3h, today], default: now }
      responses:
        "200": { description: "Demand points", content: { application/json: { schema: { $ref: "#/components/schemas/DriverHeatmapResponse" } } } }
        "400": { description: "Invalid coordinates", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
...